# mimirpb: Refactor RW2 metadata handling out of generated code