--- a/pkg/mimirpb/compat_rw2_test.go
+++ b/pkg/mimirpb/compat_rw2_test.go
@@ -759,6 +759,44 @@ func TestRW2Unmarshal(t *testing.T) {
 			require.Equal(t, "metric_2 help text.", received.Metadata[1].Help)
 		}
 	})
+
+	t.Run("conflicting metadata, first metadata wins by default", func(t *testing.T) {
+		writeRequest := &WriteRequest{
+			SymbolsRW2: []string{"", "__name__", "my_cool_series", "It's a cool series, but old description.", "It's a cool series, but new description.", "megawatts"},
+			TimeseriesRW2: []TimeSeriesRW2{
+				{
+					LabelsRefs: []uint32{1, 2},
+					Metadata: MetadataRW2{
+						Type:    METRIC_TYPE_COUNTER,
+						HelpRef: 3,
+						UnitRef: 5,
+					},
+				},
+				{
+					LabelsRefs: []uint32{1, 2},
+					Metadata: MetadataRW2{
+						Type:    METRIC_TYPE_COUNTER,
+						HelpRef: 4,
+						UnitRef: 5,
+					},
+				},
+			},
+		}
+		data, err := writeRequest.Marshal()
+		require.NoError(t, err)
+
+		// Unmarshal the data back into Mimir's WriteRequest.
+		received := PreallocWriteRequest{}
+		received.UnmarshalFromRW2 = true
+		err = received.Unmarshal(data)
+		require.NoError(t, err)
+
+		require.Len(t, received.Metadata, 1)
+		require.Equal(t, received.Metadata[0].MetricFamilyName, "my_cool_series")
+		require.Equal(t, received.Metadata[0].Type, COUNTER)
+		require.Equal(t, received.Metadata[0].Help, "It's a cool series, but old description.")
+		require.Equal(t, received.Metadata[0].Unit, "megawatts")
+	})
 }
 
 func makeTestRW2WriteRequest(syms *test.SymbolTableBuilder) *WriteRequest {