--- a/pkg/mimirpb/custom.go
+++ b/pkg/mimirpb/custom.go
@@ -452,6 +452,42 @@ type MarshalerWithSize interface {
 	MarshalWithSize(size int) ([]byte, error)
 }
 
+// metadataSet is the collection of metadata within a request.
+// It keeps the order at which metadata is added. Metadata may optionally be deduplicated by family name.
+type orderedMetadataSet struct {
+	dedupe bool
+	// We avoid unifying this into a map[string][]*Metadata or some equivalent for perf reasons.
+	// It reduces allocations and allows us to take a shortcut when not deduplicating.
+	deduplicated map[string]*orderAwareMetricMetadata
+	unmodified   []*MetricMetadata
+}
+
+func (m *orderedMetadataSet) add(family string, mm *MetricMetadata) {
+	if m.dedupe {
+		m.deduplicated[family] = &orderAwareMetricMetadata{MetricMetadata: *mm} // TODO: Make orderAware wrap the ptr and remove the inner *
+		return
+	}
+	m.unmodified = append(m.unmodified, mm)
+}
+
+func (m *orderedMetadataSet) len() int {
+	if m.dedupe {
+		return len(m.deduplicated)
+	}
+	return len(m.unmodified)
+}
+
+func (m *orderedMetadataSet) toSlice() []*MetricMetadata {
+	if !m.dedupe {
+		return m.unmodified
+	}
+	result := make([]*MetricMetadata, m.len())
+	for _, meta := range m.deduplicated {
+		result[meta.order] = &meta.MetricMetadata
+	}
+	return result
+}
+
 // orderAwareMetricMetadata is a tuple (index, metadata) that knows its own position in a metadata slice.
 // It's tied to custom logic that unmarshals RW2 metadata into a map, and allows us to
 // remember the order that metadata arrived in when unmarshalling.