--- a/pkg/mimirpb/mimir.pb.go
+++ b/pkg/mimirpb/mimir.pb.go
@@ -7385,7 +7385,7 @@ func valueToStringMimir(v interface{}) string {
 	return fmt.Sprintf("*%v", pv)
 }
 func (m *WriteRequest) Unmarshal(dAtA []byte) error {
-	var metadata map[string]*orderAwareMetricMetadata
+	var metadata metadataSet
 	seenFirstSymbol := false
 
 	l := len(dAtA)
@@ -7584,7 +7584,7 @@ func (m *WriteRequest) Unmarshal(dAtA []byte) error {
 			m.Timeseries = append(m.Timeseries, PreallocTimeseries{})
 			m.Timeseries[len(m.Timeseries)-1].skipUnmarshalingExemplars = m.skipUnmarshalingExemplars
 			if metadata == nil {
-				metadata = make(map[string]*orderAwareMetricMetadata)
+				metadata = newDedupingMetadataSet()
 			}
 			if err := m.Timeseries[len(m.Timeseries)-1].Unmarshal(dAtA[iNdEx:postIndex], &m.rw2symbols, metadata, m.skipNormalizeMetadataMetricName); err != nil {
 				return err
@@ -7651,10 +7651,7 @@ func (m *WriteRequest) Unmarshal(dAtA []byte) error {
 	}
 
 	if m.unmarshalFromRW2 {
-		m.Metadata = make([]*MetricMetadata, len(metadata))
-		for _, metadata := range metadata {
-			m.Metadata[metadata.order] = &metadata.MetricMetadata
-		}
+		m.Metadata = metadata.slice()
 		m.rw2symbols.releasePages()
 	}
 
@@ -11240,7 +11237,7 @@ func (m *WriteRequestRW2) Unmarshal(dAtA []byte) error {
 func (m *TimeSeriesRW2) Unmarshal(dAtA []byte) error {
 	return errorInternalRW2
 }
-func (m *TimeSeries) UnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols, metadata map[string]*orderAwareMetricMetadata, skipNormalizeMetricName bool) error {
+func (m *TimeSeries) UnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols, metadata metadataSet, skipNormalizeMetricName bool) error {
 	var metricName string
 	l := len(dAtA)
 	iNdEx := 0
@@ -11698,7 +11695,7 @@ func (m *Exemplar) UnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols) error {
 func (m *MetadataRW2) Unmarshal(dAtA []byte) error {
 	return errorInternalRW2
 }
-func MetricMetadataUnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols, metadata map[string]*orderAwareMetricMetadata, metricName string, skipNormalizeMetricName bool) error {
+func MetricMetadataUnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols, metadata metadataSet, metricName string, skipNormalizeMetricName bool) error {
 	var (
 		err error
 		help string
@@ -11826,22 +11823,13 @@ func MetricMetadataUnmarshalRW2(dAtA []byte, symbols *rw2PagedSymbols, metadata 
 	if len(normalizedMetricName) == 0 {
 		return nil
 	}
-	if _, ok := metadata[normalizedMetricName]; ok {
-		// Already have metadata for this metric familiy name.
-		// Since we cannot have multiple definitions of the same
-		// metric family name, we ignore this metadata.
-		return nil
-	}
 	if len(unit) > 0 || len(help) > 0 || metricType != 0 {
-		metadata[normalizedMetricName] = &orderAwareMetricMetadata{
-			MetricMetadata: MetricMetadata{
-				MetricFamilyName: normalizedMetricName,
-				Help:             help,
-				Unit:             unit,
-				Type:             MetricMetadata_MetricType(metricType),
-			},
-			order: len(metadata),
-		}
+		metadata.add(normalizedMetricName, MetricMetadata{
+			MetricFamilyName: normalizedMetricName,
+			Help:             help,
+			Unit:             unit,
+			Type:             MetricMetadata_MetricType(metricType),
+		})
 	}
 
 	return nil