#### What this PR does

The RW2 unmarshal uses a `map[string]*orderAwareMetricMetadata` that's injected everywhere, used to track and deduplicate metadata, and also store them as we convert back to RW1.

This PR is a refactor that puts this behind a `metadataSet` interface. We then introduce a `dedupingMetadataSet` implementation that just wraps the very same map. In the future we want to introduce alternate metadata handling strategies that don't deduplicate, this makes it easy.

Just converting it to a `map[string][]*orderAwareMetricMetadata` is too unwieldy and allocations inside the inner slices are difficult to control, this approach is more minimal. It also means we can move a little custom logic out of `mimir.pb.go` and into regular Go files, reducing the divergence of our edits to the generated code.

This abstraction doesn't cause any perf difference or extra allocations.
<details>

<summary>benchmark results</summary>

Run with:
```
go test ./pkg/mimirpb/... -run ^$ -bench BenchmarkUnMarshal -benchtime 5s -benchmem
```

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/mimirpb
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics
                                               │  before.txt  │              after.txt               │
                                               │    sec/op    │    sec/op     vs base                │
UnMarshal/Marshal/RW1-16                         39.04m ± ∞ ¹   38.45m ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Marshal/RW2-16                         21.19m ± ∞ ¹   21.08m ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW1/skipExemplars=true-16    24.75m ± ∞ ¹   23.54m ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=true-16    26.15m ± ∞ ¹   26.80m ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW1/skipExemplars=false-16   27.58m ± ∞ ¹   26.16m ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=false-16   28.39m ± ∞ ¹   29.15m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                          27.36m         27.03m        -1.20%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                               │  before.txt   │               after.txt               │
                                               │     B/op      │     B/op       vs base                │
UnMarshal/Marshal/RW1-16                         25.58Mi ± ∞ ¹   25.58Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Marshal/RW2-16                         17.50Mi ± ∞ ¹   17.50Mi ± ∞ ¹       ~ (p=1.000 n=1) ³
UnMarshal/Unmarshal/RW1/skipExemplars=true-16    54.48Mi ± ∞ ¹   54.48Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=true-16    31.28Mi ± ∞ ¹   31.28Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW1/skipExemplars=false-16   59.05Mi ± ∞ ¹   59.05Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=false-16   32.80Mi ± ∞ ¹   32.80Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                          33.75Mi         33.75Mi        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal

                                               │  before.txt  │              after.txt               │
                                               │  allocs/op   │  allocs/op    vs base                │
UnMarshal/Marshal/RW1-16                          2.000 ± ∞ ¹    2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Marshal/RW2-16                          2.000 ± ∞ ¹    2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW1/skipExemplars=true-16    60.43k ± ∞ ¹   60.43k ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=true-16    50.02k ± ∞ ¹   50.02k ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW1/skipExemplars=false-16   100.4k ± ∞ ¹   100.4k ± ∞ ¹       ~ (p=1.000 n=1) ²
UnMarshal/Unmarshal/RW2/skipExemplars=false-16   60.02k ± ∞ ¹   60.02k ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                          2.044k         2.044k        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```

<details>

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.